requirements.txt
google-genai
fastapi
uvicorn[standard]
websockets
httpx
python-dotenv
numpy
pyngrok
